WARNING! Might have problems!
ActiveLed doesn't behave correctly with OneShot here. Specifically it
doesn't go on and off....eventually the OneShot keys will light up and
stay lit up. I'm not sure if that's because of a bug in THIS firmware or
the plugin bundle I built it against.

I'm creating this commit because it also brings this firmware up to
being source-compatible with the latest bundle...specifically functions
on OneShot instead of fields.